Allow 'exact' match to be enforced when selecting radios
We change the default behaviour to be inexact in order to
make our tests more resilient to change. The copy on radio
buttons will change, and that ought not to break our
acceptance tests.

In this commit we set one checkbox action to an 'exact' match
in order to avoid an error where multiple options were
matched:

```
locator.check: Error: strict mode violation: getByLabel('No') resolved to 2 elements:
    1) <input value="no" type="radio" id="hasDisability-2" nam…/> aka getByLabel('No', { exact: true })
    2) <input type="radio" id="hasDisability-4" name="hasDisab…/> aka getByLabel('Prefer not to say')

=========================== logs ===========================
waiting for getByLabel('No')
============================================================

```