Namespace CircleCI variables to avoid conflicts
When adding e2e tests to the API pipeline we have to set environment variables alongside other the CAS projects. Sadly `HMPPS_AUTH_USERNAME` was already defined and with different values so we have to distinguish them. Given we have to prefix one we decide to prefix all just for improved clarity from the API context.

Instead of trying to allow the old names on github actions and in local .env files we take the step to change this everywhere.